We made some progress on the quiz window site since the hint and answers are now being displayed.
but not yet properly and in their own css containers

but i was able to make the remaining options radiobuttons, apply the css classes to them and give them the correct alphabetical letters

AND EVEN BETTER, we were able to fix the embedded text classes not properly applying.

and at the same time also fixed the bold class not being properly applied aswell. Well done!