Daily backup
Made optimizations for the student dashboards styling and added logic for managing the amound of quizzes displayed on the dasboard.

added pagnation to the student dashboard.

A temporary global header for all pages that displays page redirects links based on the loged in users user role

SOme changes to the dashboard for admins and editors